Structural pattern: Composite pattern
