+++ app/location/nyt.py
@@ -1,5 +1,6 @@
 """app.locations.nyt.py"""
 from . import TimelinedLocation
+from . import serialize
 
 
 class NYTLocation(TimelinedLocation):
@@ -14,19 +15,4 @@ class NYTLocation(TimelinedLocation):
         self.state = state
         self.county = county
 
-    def serialize(self, timelines=False):  # pylint: disable=arguments-differ,unused-argument
-        """
-        Serializes the location into a dict.
-
-        :returns: The serialized location.
-        :rtype: dict
-        """
-        serialized = super().serialize(timelines)
-
-        # Update with new fields.
-        serialized.update(
-            {"state": self.state, "county": self.county,}
-        )
-
-        # Return the serialized location.
-        return serialized
+
